requirements.txt
anthropic==0.37.1
google-generativeai==0.8.3
openai==1.52.2
pandas==2.2.3
python-dotenv==1.0.1
tqdm==4.66.5
groq==0.19.0